Extract public streaming session types
